First Milestone Met
My first milestone isn't quite where I wanted to be by this time, however this is mostly down to a busy week and grappling with a bizarre Godot bug that kept failing to build my project and caused me to carpetbomb a bunch of progress in a frenzied effort to solve it. That being said, I'm pretty happy with how the crowds are turning out on the whole- leader crowds aren't a thing yet, however I don't think they'll be difficult to implement within the next couple days. The toughest part, by far, was figuring out exactly how Godot deals with basic OOP stuff, and how I could get initialization scripts to run when I instanced pawns to populate crowds.