Final Commit! Version 1.0!!
The game is now in a playable state! It's changed a lot over the course of development, and a lot of my stretch goals were, as I found, less unatainable and more simply distracting from the focus of the game- so I got rid of them for simplicity's sake. I'm very happy with where the game is right now